requirements.txt
streamlit>=1.32.0
google-cloud-storage>=2.14.0
google-genai>=1.13.0
python-dotenv>=1.0.1
Pillow>=10.0.0 